// frontend/src/app/api/model/addressDetailsDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddressDetailsDto { 
    cityId: number;
    street: string;
    district?: string;
    building?: string;
    entrance?: string;
    floor?: string;
    apartment?: string;
    latitude?: number;
    longitude?: number;
}